Move userFriendlyErrors to utils
